+++ src/model/graphics/document_scene.cpp
@@ -19,7 +19,7 @@ void model::graphics::DocumentScene::set_document ( model::Document* document )
 {
     if ( d->document )
     {
-        disconnect_node(&d->document->animation());
+        disconnect_node(d->document->animation());
     }
 
     clear();
@@ -28,7 +28,7 @@ void model::graphics::DocumentScene::set_document ( model::Document* document )
 
     if ( d->document )
     {
-        connect_node(&d->document->animation());
+        connect_node(d->document->animation());
     }
 }