+++ src/scripting/js/js_engine.cpp
@@ -1,0 +1,4 @@
+#include "js_engine.hpp"
+
+
+scripting::ScriptEngine::Autoregister<scripting::js::JsEngine> scripting::js::JsEngine::autoreg;