+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Unreleased
  * Layer Management
  * Open/Save Glaxnim files
+ * JS Scripting capabilities